# expression: implement vectorized evaluation for `builtinStrToDateDurationSig`