--- a/expression/builtin_time_vec_test.go
+++ b/expression/builtin_time_vec_test.go
@@ -280,6 +280,12 @@ var vecBuiltinTimeCases = map[string][]vecExprBenchCase{
 			// "%y%m%d" is wrong format, STR_TO_DATE should be failed for all rows
 			constants: []*Constant{nil, {Value: types.NewDatum("%y%m%d"), RetType: types.NewFieldType(mysql.TypeString)}},
 		},
+		{
+			retEvalType:   types.ETDuration,
+			childrenTypes: []types.EvalType{types.ETString, types.ETString},
+			geners:        []dataGenerator{&dateStrGener{}, nil},
+			constants:     []*Constant{nil, {Value: types.NewDatum("%H:%i:%s"), RetType: types.NewFieldType(mysql.TypeString)}},
+		},
 	},
 	ast.GetFormat: {
 		{