PCP #12176 

### What problem does this PR solve? <!--add issue link with summary if exists-->
Implement vectorized evaluation for `builtinStrToDateDurationSig` at #12176 

### What is changed and how it works?
It gets faster about 28.7%.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
```
> GO111MODULE=on go test -check.f TestVectorizedBuiltinTimeFunc
PASS: builtin_time_vec_test.go:239: testEvaluatorSuite.TestVectorizedBuiltinTimeFunc	0.157s
PASS: builtin_time_vec_generated_test.go:192: testEvaluatorSuite.TestVectorizedBuiltinTimeFuncGenerated	0.110s
OK: 2 passed
PASS
ok  	github.com/pingcap/tidb/expression	0.292s
```

 - Benchmark
```
> go test -v -benchmem -bench=BenchmarkVectorizedBuiltinTimeFunc -run=BenchmarkVectorizedBuiltinTimeFunc -args builtinStrToDateDurationSig
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinTimeFuncGenerated-12    	1000000000	         0.0165 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinStrToDateDurationSig-VecBuiltinFunc-12         	    1444	    867137 ns/op	  284975 B/op	    3073 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinStrToDateDurationSig-NonVecBuiltinFunc-12      	    1122	    901480 ns/op	  285244 B/op	    3073 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	2.644s
```